requirements.txt
requests==2.32.3
Flask==2.2.5
oauthlib==3.0.1
pyOpenSSL==24.2.1
Flask-Login==0.6.3
python-dotenv==1.0.1
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
gunicorn==22.0.0
